+++ aics-admin/src/main/java/kgu/developers/admin/post/presentation/request/PostRequest.java
@@ -1,18 +1,20 @@
-package kgu.developers.api.post.presentation.request;
+package kgu.developers.admin.post.presentation.request;
 
-import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
+import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.*;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
+import kgu.developers.domain.post.domain.Category;
 import lombok.Builder;
 
 @Builder
 public record PostRequest(
+
 	@Schema(description = "게시물 제목", example = "SW 부트캠프 4기 교육생 모집", requiredMode = REQUIRED)
-	@NotBlank
-	@Size(max = 100, message = "제목은 100자 이내여야 합니다.")
+	@NotBlank @Size(max = 100, message = "제목은 100자 이내여야 합니다.")
 	String title,
+
 	@Schema(description = "게시물 내용",
 		example = """
 			SW 부트캠프 4기 교육생 모집
@@ -23,6 +25,9 @@ public record PostRequest(
 			""",
 		requiredMode = Schema.RequiredMode.REQUIRED)
 	@NotBlank
-	String content
+	String content,
+
+	@Schema(description = "게시물 카테고리", example = "DEPT_INFO", requiredMode = NOT_REQUIRED)
+	Category category
 ) {
 }
