+++ aics-api/src/main/java/kgu/developers/api/post/presentation/PostController.java
@@ -1,19 +1,13 @@
 package kgu.developers.api.post.presentation;
 
-import static org.springframework.http.HttpStatus.CREATED;
-
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import io.swagger.v3.oas.annotations.Hidden;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -22,10 +16,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.PositiveOrZero;
-import kgu.developers.api.post.application.PostService;
-import kgu.developers.api.post.presentation.request.PostRequest;
-import kgu.developers.api.post.presentation.response.PostDetailResponse;
-import kgu.developers.api.post.presentation.response.PostPersistResponse;
+import kgu.developers.api.post.application.PostFacade;
+import kgu.developers.domain.post.application.response.PostDetailResponse;
 import kgu.developers.api.post.presentation.response.PostSummaryPageResponse;
 import kgu.developers.domain.post.domain.Category;
 import lombok.RequiredArgsConstructor;
@@ -35,21 +27,7 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("/api/v1/posts")
 @Tag(name = "Post", description = "게시글 API")
 public class PostController {
-	private final PostService postService;
-
-	@Operation(summary = "게시글 생성 API", description = """
-			- Description : 이 API는 게시글을 생성합니다.
-			- Assignee : 박민준
-		""")
-	@ApiResponse(responseCode = "201", content = @Content(schema = @Schema(implementation = PostPersistResponse.class)))
-	@PostMapping
-	public ResponseEntity<PostPersistResponse> createPost(
-		@Parameter(description = "게시글 카테고리", example = "DEPT_INFO", required = true) @RequestParam Category category,
-		@RequestBody PostRequest request
-	) {
-		PostPersistResponse response = postService.createPost(request, category);
-		return ResponseEntity.status(CREATED).body(response);
-	}
+	private final PostFacade postFacade;
 
 	@Operation(summary = "게시글 페이징 조회 API", description = """
 		    - Description : 이 API는 키워드를 이용하여 게시글을 페이징 조회합니다.
@@ -63,7 +41,7 @@ public class PostController {
 		@Parameter(description = "응답 개수", example = "10", required = true) @RequestParam(defaultValue = "10") @Positive int size,
 		@Parameter(description = "검색 키워드", example = "컴퓨터공학과") @RequestParam(required = false) String keyword
 	) {
-		PostSummaryPageResponse response = postService.getPostsByKeywordAndCategory(PageRequest.of(page, size), keyword,
+		PostSummaryPageResponse response = postFacade.getPostsByKeywordAndCategory(PageRequest.of(page, size), keyword,
 			category);
 		return ResponseEntity.ok(response);
 	}
@@ -77,54 +55,7 @@ public class PostController {
 	public ResponseEntity<PostDetailResponse> getPostById(
 		@Parameter(description = "조회할 게시글의 id", example = "1", required = true) @PathVariable @Positive Long postId
 	) {
-		PostDetailResponse response = postService.getPostByIdWithPrevAndNext(postId);
-		return ResponseEntity.ok(response);
-	}
-
-	@Operation(summary = "게시글 수정 API", description = """
-		    - Description : 이 API는 게시글을 수정합니다.
-		    - Assignee : 박민준
-		""")
-	@ApiResponse(responseCode = "204")
-	@PatchMapping("/{postId}")
-	public ResponseEntity<Void> updatePost(
-		@Parameter(description = "수정할 게시글의 id", example = "1", required = true) @PathVariable @Positive Long postId,
-		@RequestBody PostRequest request
-	) {
-		postService.updatePost(postId, request);
-		return ResponseEntity.noContent().build();
-	}
-
-	@Operation(summary = "게시글 상단 고정 상태 토글 API", description = """
-		    - Description : 이 API는 지정된 게시글의 고정 여부를 토글하여 고정 또는 해제합니다.
-		    - Assignee : 박민준
-		""")
-	@ApiResponse(responseCode = "204")
-	@PatchMapping("/{postId}/pin")
-	public ResponseEntity<Void> togglePostPinStatus(
-		@Parameter(description = "고정 상태를 변경할 게시글의 id", example = "1", required = true) @PathVariable @Positive Long postId
-	) {
-		postService.togglePostPinStatus(postId);
-		return ResponseEntity.noContent().build();
-	}
-
-	@Operation(summary = "게시글 삭제 API", description = """
-		    - Description : 이 API는 게시글을 삭제합니다.
-		    - Assignee : 이신행
-		""")
-	@ApiResponse(responseCode = "204")
-	@PatchMapping("/{postId}/delete")
-	public ResponseEntity<Void> deletePostById(
-		@Parameter(description = "조회할 게시글의 id", example = "1", required = true) @PathVariable @Positive Long postId
-	) {
-		postService.deletePost(postId);
-		return ResponseEntity.noContent().build();
-	}
-
-	@Hidden
-	@GetMapping("/cleanup-last-run")
-	public ResponseEntity<String> getLastCleanupRunTime() {
-		String response = postService.getFormattedLastCleanupRunTime();
+		PostDetailResponse response = postFacade.getPostByIdWithPrevAndNext(postId);
 		return ResponseEntity.ok(response);
 	}
 }
