+++ aics-api/src/testFixtures/java/mock/TestContainer.java
@@ -5,7 +5,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 import kgu.developers.api.auth.application.AuthService;
 import kgu.developers.api.auth.presentation.AuthController;
-import kgu.developers.api.post.application.PostService;
+import kgu.developers.api.post.application.PostFacade;
 import kgu.developers.api.user.application.UserFacade;
 import kgu.developers.common.auth.jwt.JwtProperties;
 import kgu.developers.common.auth.jwt.TokenProvider;
@@ -14,10 +14,10 @@ import kgu.developers.domain.user.domain.UserRepository;
 
 public class TestContainer {
 	public final UserRepository userRepository;
-	public final UserFacade userFacade;
+	// public final UserFacade userFacade;
 	public final AuthService authService;
 	public final AuthController authController;
-	public final PostService postService;
+	// public final PostFacade postFacade;
 	public final PostRepository postRepository;
 	public final RefreshTokenRepository refreshTokenRepository;
 
@@ -25,12 +25,12 @@ public class TestContainer {
 		BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
 		this.refreshTokenRepository = new FakeRefreshTokenRepository();
 		this.userRepository = new FakeUserRepository();
-		this.userFacade = UserFacade.builder()
-			.userRepository(this.userRepository)
-			.bCryptPasswordEncoder(bCryptPasswordEncoder)
-			.build();
+		// this.userFacade = UserFacade.builder()
+			// .userRepository(this.userRepository)
+			// .bCryptPasswordEncoder(bCryptPasswordEncoder)
+			// .build();
 		this.authService = AuthService.builder()
-			.userFacade(this.userFacade)
+			// .userFacade(this.userFacade)
 			.passwordEncoder(bCryptPasswordEncoder)
 			.tokenProvider(TokenProvider.builder()
 				.jwtProperties(new JwtProperties("testIssuer", "testSecretKey"))
@@ -43,6 +43,6 @@ public class TestContainer {
 			.build();
 
 		this.postRepository = new FakePostRepository();
-		this.postService = new PostService(postRepository, userFacade);
+		// this.postFacade = new PostFacade(postRepository, userFacade);
 	}
 }
