+++ aics-domain/src/main/java/kgu/developers/domain/post/application/response/PostTitleResponse.java
@@ -1,4 +1,4 @@
-package kgu.developers.api.post.presentation.response;
+package kgu.developers.domain.post.application.response;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
