+++ aics-domain/src/main/java/kgu/developers/domain/post/exception/PostNotFoundException.java
@@ -1,0 +1,11 @@
+package kgu.developers.domain.post.exception;
+
+import static kgu.developers.domain.post.exception.PostDomainExceptionCode.*;
+
+import kgu.developers.common.exception.CustomException;
+
+public class PostNotFoundException extends CustomException {
+	public PostNotFoundException() {
+		super(POST_NOT_FOUND);
+	}
+}
